Fix Windows build by excluding fuzz targets from workspace build
The fuzz crate uses #![no_main] which causes linker errors on Windows
when built with `cargo build`. While Linux/Mac linkers seems to be permissive and allow this, the resulting binaries are probably non-functional. It seems fuzz targets should only be built with `cargo fuzz` as the directory README explain